Don't Force Power Stars to be Local in SM64
It doesn't seem like Figment is going to update their pull request any time soon, so I'll bring this out there again.

This has been discussed and the consensus seems to be that Power Stars in pool will work out better.

If Power Stars are local, the majority of a player's items are going to be in their own world; they will complete all of Floor 1 in their first log-in; and they will be sending very few items into the multiworld. If Power Stars are in the pool, they will log back in multiple times to complete the first floor; and players will be sending out items, allowing them to feel like they are part of the multiworld and not playing a mostly solo game.