SA2: Add new options
I don't feel strongly about the junk fill and trap fill, or whether traps should be included at all; I could see changing random-low to a random-range with lower values as well, but also don't feel particularly strongly about it. Anyone else is free to make that suggestion or change though.

I wanted a 50/50 chance of Chao Garden, with a leaning towards more of it if it gets put in; Chao Karate less so, because it requires the DLC. So including whether Chao Karate is on or off should be expected on the spreadsheet.

Setting rank to E to encourage new players to take these worlds; an experienced player will probably be making a custom. Could still see this change though.