Update to 0.3.7.
My rationale behind these choices:

Chaos Emeralds can be a bit of a slog and Triforce Hunt-adjacent (more so than Emblems are), so let's leave it out.

Chao Keys and Pipes are doable (and there's a good guide out), but let's keep the chance of them lower. Beetles in general are very easy, but keeping the same random weights for consistency.

For level types, a little bit of variety is good. Keeping Chao missions (Mission 3) in the rotation ensures the Mystic Melodies have a chance to be required, and otherwise the missions are not hard. Ring Missions (Mission 2) are very easy. Speed/Mech Mission 4s (Timer) are generally interesting and worth having a chance to show up.

Speed Missions randomly have 2-4 missions each because Speed missions are generally the most fun and interesting. If it rolls 2 or 3, it guarantees there will be variety in the mission type seen; if it rolls 4, at least they will be in random order.

All other missions are 1 or 2, with some variety to allow different mission types to be experienced.

Kart Mission 2 are extremely easy and if Mission Count is rolled 1, the chance that people will only have to do Rings and not the normal run would be a relief.

I haven't actually rolled this yet to confirm it works; will do that closer to the async time. I'm playing in another async and I don't want to update to 0.3.7 in case something breaks. But I want to propose this here both so I don't forget about it and to open discussion.