Update Super Mario 64.yaml
Updated YAML with new options or options that had their names changed.

Options that were previously weighted remain at that same weight with the exception of `stars_to_finish` since that is now percentage-based. It now uses a `random-range-middle` with the range starting and ending at percentages approximately equal to the previous maximum and minimum weighted options.

Any new options added that are weighted towards a particular setting use the same weights as the previous version (5-to-2) to favor what is generally considered the easier or more desirable option of the two.