Add game: A Hat in Time
AHit has a lot of options and many of them interact with other options.
The interactions between most options have been controlled using
triggers, which makes the yaml quite complicated.

The main important option is the trigger for controlling which DLCs
are enabled and thus required for the player to own. I could not find
any charts showing the DLC ownership percentages, so I took a guess at
what might be common for the AP community.
The most common DLC1 achievement is Prepare to Die at 12.6% of players,
and the most common DLC2 achievement is Stick It To The Man at 16.3% of
players. I anticipate that DLC ownership is higher (especially in the
case of DLC1 because the achievement is more difficult) and the DLCs are
fairly low in price compared to the game itself.

For filler-only asyncs, there is a trigger that allows harder options to
be enabled. Disabling this will replace all harder options with easier
ones towards the end of the triggers.