Adds Chao Stats
Redid the weights for Chao Race, and moved Chao Karate to be triggered based off of what Chao Race rolls. They are more likely to be similar difficulty but there is still a chance that they can be mis-matched. Also in the triggers are chao stat settings that set the max check based off of the highest level of chao stadium checks. For example, if beginner is rolled for both, the max chao level check is 20. If race has Intermediate but Karate has Super, the max level is 70. All options have equal weight of being a check per 1, 2, or 5 levels.