add trigger for Any goal.yaml
Due to unnoticed behaviour regarding The "Any" option for goal, the choice was either to create a second trigger for the Any goal to make grubs local, or to remove Any as an option. I feel as if I saw enough positive opinions on Any goal being like this that I went with the former.