Add fast_travel_only start inventory option at 15%
`fast_travel_only` gives only the `Wind Waker` and `Ballad of Gales` as
starting items and is only available in filler-only asyncs.

Originally I was going to go with a 5-10% chance, but it seems like
starting with only Wind Waker + Ballad might be more popular than I
thought, so I have gone with a 15% chance.

This reduces the chance of `basics` by 15%.

Co-authored-by: Jonathan Tan <tanjo3@users.noreply.github.com>