Reduce spoils trading and big octos+gunboats chances
These options tend to be more tedious and repetitive.

Eye reefs almost all become in-logic at the same time and tend to be
accessed later in most seeds, requiring Wind Waker + Wind's Requiem +
Deku Leaf + Progressive Magic Meter + Bombs (or Boomerang in one case).

Big Octos mostly become in-logic at the same time, the differences being
Big Octos with many eyes requiring a Progressive Quiver Upgrade if the
player has a Bow, but no Bombs/Boomerang and the one Big Octo that is
also a Great Fairy which does not require Grappling Hook.

The main differences to me are that Big Octos have a presence in the
game earlier on and have more variety in the items that can be used to
defeat them. Disabling Big Octo locations does not disable the Big Octos
themselves, so Big Octos are sometimes encountered by accident while
sailing around the sea, though this should be less common when starting
with Wind Waker + Ballad of Gales.

Eye reefs chance reduced from 80% to 50%
Big Octos/Gunboats chance reduced from 90% to 60%

Co-authored-by: Jonathan Tan <tanjo3@users.noreply.github.com>