Decrease expensive purchases chance from 85% to 65%
Expensive purchases covers Auctions, Mail from Tingle, Triforce Charts,
the full Goron Trading sidequest, and the expensive Rock Spire Isle
shop.

Auctions, the Mail from Tingle, and the Goron Trading quest are not that
bad in terms of required Rupees, and the Auctions tell you what item
they have.

Triforce Charts can be costly, but might not even be enabled. In
auto-collect Big Asyncs, the player may not need to do all of these
checks either.
Auctions are also considered Minigames, so might not be enabled.
The Goron Trading sidequest is also considered a Long Sidequest, so
is unlikely to be enabled.
The Rock Spire Isle shop is costly, but the shop shows what the items
are. Savescumming can be used to avoid having to acquire the full amount
of Rupees to buy all 3 items, but I don't think this should be an
assumed strategy of TWW players.

Co-authored-by: Jonathan Tan <tanjo3@users.noreply.github.com>