Decrease platforms+rafts chance from 90% to 70%
Platforms + rafts add a lot of locations, a bunch of which are sphere 1,
but while the locations overall can be quite tedious, it is broken up
into more location checks than other tedious options, so I don't think
it needs to be as low of a chance as the other tedious options.

Big Asyncs are always generated without progression balancing, so while
logic may dictate that some progression items have to be placed in early
spheres for the multiworld to progress, there should not be any extra
bias towards early spheres like would be present in multiworlds with
progression balancing enabled, so the extra sphere 1 locations should
not be a big issue.

1 in 6 slots will generate with platforms+rafts forcefully enabled as
the baseline location source, so the individual chance does not need to
be too high.

Co-authored-by: Jonathan Tan <tanjo3@users.noreply.github.com>