Decrease treasure/triforce charts chance from 90% to 65%
Treasure charts already have a 2/6 chance to be picked as the baseline
locations, so the 90% chance was too high.

Reducing to 65% should give around a 77% chance that some charts are
enabled, though its actually slightly lower because `triforce_only` can
result in no charts being enabled when expensive purchases are disabled.

The order of the `trigger_progression_charts` option weights has been
flipped to better match the rest of the yaml, with false/disabled first.

Co-authored-by: Jonathan Tan <tanjo3@users.noreply.github.com>