Add The Wind Waker.yaml (#267)
* Add The Wind Waker.yaml

* Simplify disabling of secret cave ER when either type are disabled

* Normalize start inventory weights for non-filler asyncs

* Remove duplicated dungeon/miniboss/boss ER cleanup

* Apply suggestions from code review

Adjusted start inventory weights, greatly decreasing `nothing` chance and decreasing `base_rando` chance
Decreased long sidequests chance
Decreased spoils trading chance
Maps and compasses are now always started with
Increased ER chance
Increased the chances off ER modes other than dungeons/nested dungeons
`reveal_full_sea_chart` is now always enabled
Asyncs that allow custom submissions are now allowed to roll dungeon secrets and/or tingle chests

Co-authored-by: Jonathan Tan <tanjo3@users.noreply.github.com>

* Comment updates

* Add fast_travel_only start inventory option at 15%

`fast_travel_only` gives only the `Wind Waker` and `Ballad of Gales` as
starting items and is only available in filler-only asyncs.

Originally I was going to go with a 5-10% chance, but it seems like
starting with only Wind Waker + Ballad might be more popular than I
thought, so I have gone with a 15% chance.

This reduces the chance of `basics` by 15%.

Co-authored-by: Jonathan Tan <tanjo3@users.noreply.github.com>

* Only allow non-skipped rematch bosses when dungeons are disabled

Co-authored-by: Jonathan Tan <tanjo3@users.noreply.github.com>

* Reduce spoils trading and big octos+gunboats chances

These options tend to be more tedious and repetitive.

Eye reefs almost all become in-logic at the same time and tend to be
accessed later in most seeds, requiring Wind Waker + Wind's Requiem +
Deku Leaf + Progressive Magic Meter + Bombs (or Boomerang in one case).

Big Octos mostly become in-logic at the same time, the differences being
Big Octos with many eyes requiring a Progressive Quiver Upgrade if the
player has a Bow, but no Bombs/Boomerang and the one Big Octo that is
also a Great Fairy which does not require Grappling Hook.

The main differences to me are that Big Octos have a presence in the
game earlier on and have more variety in the items that can be used to
defeat them. Disabling Big Octo locations does not disable the Big Octos
themselves, so Big Octos are sometimes encountered by accident while
sailing around the sea, though this should be less common when starting
with Wind Waker + Ballad of Gales.

Eye reefs chance reduced from 80% to 50%
Big Octos/Gunboats chance reduced from 90% to 60%

Co-authored-by: Jonathan Tan <tanjo3@users.noreply.github.com>

* Decrease expensive purchases chance from 85% to 65%

Expensive purchases covers Auctions, Mail from Tingle, Triforce Charts,
the full Goron Trading sidequest, and the expensive Rock Spire Isle
shop.

Auctions, the Mail from Tingle, and the Goron Trading quest are not that
bad in terms of required Rupees, and the Auctions tell you what item
they have.

Triforce Charts can be costly, but might not even be enabled. In
auto-collect Big Asyncs, the player may not need to do all of these
checks either.
Auctions are also considered Minigames, so might not be enabled.
The Goron Trading sidequest is also considered a Long Sidequest, so
is unlikely to be enabled.
The Rock Spire Isle shop is costly, but the shop shows what the items
are. Savescumming can be used to avoid having to acquire the full amount
of Rupees to buy all 3 items, but I don't think this should be an
assumed strategy of TWW players.

Co-authored-by: Jonathan Tan <tanjo3@users.noreply.github.com>

* Decrease platforms+rafts chance from 90% to 70%

Platforms + rafts add a lot of locations, a bunch of which are sphere 1,
but while the locations overall can be quite tedious, it is broken up
into more location checks than other tedious options, so I don't think
it needs to be as low of a chance as the other tedious options.

Big Asyncs are always generated without progression balancing, so while
logic may dictate that some progression items have to be placed in early
spheres for the multiworld to progress, there should not be any extra
bias towards early spheres like would be present in multiworlds with
progression balancing enabled, so the extra sphere 1 locations should
not be a big issue.

1 in 6 slots will generate with platforms+rafts forcefully enabled as
the baseline location source, so the individual chance does not need to
be too high.

Co-authored-by: Jonathan Tan <tanjo3@users.noreply.github.com>

* Decrease treasure/triforce charts chance from 90% to 65%

Treasure charts already have a 2/6 chance to be picked as the baseline
locations, so the 90% chance was too high.

Reducing to 65% should give around a 77% chance that some charts are
enabled, though its actually slightly lower because `triforce_only` can
result in no charts being enabled when expensive purchases are disabled.

The order of the `trigger_progression_charts` option weights has been
flipped to better match the rest of the yaml, with false/disabled first.

Co-authored-by: Jonathan Tan <tanjo3@users.noreply.github.com>

* Decrease randomize_charts chance from 25% to 10%

It does not add much to the randomizer and the lack of support by the
PopTracker pack may be confusing to newer players.

The base TWW rando tracker does support randomized charts but expects
the player to start with some items they might not start with in AP TWW.

Co-authored-by: Jonathan Tan <tanjo3@users.noreply.github.com>

* Add trigger for rolling an average of X random ER categories

The weight of each X can be adjusted separately for finer control, but
these are only averages, so each trigger value can still end up rolling
any number of enabled ER categories.

* Increase average random ER category count

Because disabled dungeons/secret caves/fairy fountains will disable
the corresponding ER categories, the average ER category count was lower
than expected.

* Reroll secret caves weights with more preference to both for ER

When ER is enabled and has rolled one of the random_ or all_ modes,
reroll secret caves weights with an increased preference towards both
being enabled, because secret cave ER is only enabled when both secret
caves types are enabled.

Without this, secret cave ER was enabled too infrequently for my
liking.

---------

Co-authored-by: Jonathan Tan <tanjo3@users.noreply.github.com>