Fix mission distribution
When someone updated the yaml to match 0.6.1 they did not update the missions properly. I personally submitted a custom SA2B last time so I didn't notice. This has allowed missions that were not enabled in previous Big Asyncs to be turned on. I have adjusted the yaml to be in line with what the previous settings were supposed to reflect