[Civ VI] Excluded Boost Space Race and Urbanization (#297)
* Excluded Boost Space Race and urbanization

Boosts don't have an era set they should be achievable by. 
So logic assumes these can be completed in the ancient era which they can't be.

* 2 More

2 More problematic Boosts missing requirements in Prog Era + Boosts

* Excluded 7 Boosts total

These locations are missing requirements to do them in game
A PR for a fix is open here: https://github.com/ArchipelagoMW/Archipelago/pull/5296