Create custom override for month sensitive options (#310)
This change is an evolution of the previous "is it april" option that is now automated based on the month AsyncTools is run, as opposed to changing it manually.

Note that, if it is still desired to change the options manually, it can be done so by changing the trigger option result.