[TUNIC] Update to 0.6.4
Updates TUNIC yaml to 0.6.4, as of RC1.

The only two options added are shuffle_bells and shuffle_fuses. bells I decided was minor enough that it might as well be on in every yaml. Fuses is a bit more major, and so while I assume there are very few people who would prefer it off, I think allowing some with it off would be fine